refactor : GeneralLottoNumber 패키지 이동
