+++ src/test/kotlin/lottogame/model/LottoTest.kt
@@ -1,6 +1,5 @@
 package lottogame.model
 
-import lottogame.model.generator.GeneralLottoNumber
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
